lazytest: Runs Only Failed Tests

A companion package to the testthat package that saves development time by helping you rerun only the tests that have failed during the last run.

See also

Author

Maintainer: Kirill Müller kirill@cynkra.com (ORCID)

Other contributors:

cynkra GmbH mail@cynkra.com [funder, copyright holder]